Migrate demo to `svelte-kit`
