--- src/routes/index.html
@@ -1,55 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width,initial-scale=1" />
-    <title>svelte-toast | DEMO</title>
-    <meta
-      name="description"
-      content="Simple elegant toast notifications. Use it in Vanilla JS or as a Svelte component."
-    />
-    <script>
-      window.dataLayer = window.dataLayer || []
-      function gtag() {
-        window.dataLayer.push(arguments)
-      }
-      gtag('js', new Date())
-      gtag('config', 'G-G9JC5N7N1H')
-      location.protocol === 'https:' &&
-        (function () {
-          const script = document.createElement('script')
-          script.async = true
-          script.src = 'https://www.googletagmanager.com/gtag/js?id=G-G9JC5N7N1H'
-          document.head.appendChild(script)
-        })()
-    </script>
-    <link
-      href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=swap"
-      rel="stylesheet"
-    />
-    <link rel="stylesheet" href="build/global.css" />
-    <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/prismjs@1/themes/prism-tomorrow.min.css"
-    />
-    <script defer src="build/bundle.js"></script>
-  </head>
-
-  <body class="pt-8 pb-16">
-    <h1 class="text-4xl text-center">svelte-toast</h1>
-    <div class="h-16 flex flex-col items-center">
-      <a href="https://github.com/zerodevx/svelte-toast" title="Fork me on Github">
-        <img
-          class="opacity-97 hover:opacity-75"
-          src="https://img.shields.io/npm/v/@zerodevx/svelte-toast/latest?logo=github&style=for-the-badge&color=blue&label=github"
-          alt="@zerodevx"
-        />
-      </a>
-    </div>
-    <p class="container px-2 mb-4">
-      Simple, elegant, generic, flexible and performant toast notifications. Because a demo helps
-      better than a thousand API docs, so here it is. Use in Vanilla JS (8kb gzipped) or as a Svelte
-      component.
-    </p>
-  </body>
-</html>
