--- src/routes/main.js
@@ -1,5 +1,0 @@
-import App from './App.svelte'
-
-const app = new App({ target: document.body })
-
-export default app
