--- src/routes/tailwind.css
@@ -1,25 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@layer base {
-  html {
-    @apply h-full bg-gray-50 text-gray-900 font-sans font-normal;
-    -webkit-tap-highlight-color: transparent;
-  }
-  a {
-    @apply text-blue-400 hover:text-blue-600 underline;
-  }
-  button {
-    @apply w-40 h-12 text-gray-600 hover:text-gray-900 text-xs font-bold tracking-wide bg-white border border-gray-300 hover:border-blue-800 rounded-sm focus:outline-none focus:ring px-2 m-2;
-  }
-  button.selected {
-    @apply bg-blue-200;
-  }
-}
-
-@layer components {
-  .opacity-97 {
-    opacity: 0.97;
-  }
-}
